[Hot-Fix] Fix regression in CL based codes

+++ src/imgaug/cl/cl_contrast_stretch.cpp
@@ -30,11 +30,11 @@ cl_contrast_stretch (    cl_mem srcPtr, RppiSize srcSize,
     clSetKernelArg(theKernel, 8, sizeof(unsigned int), &channel);
     //-----
 
-    unsigned int gDim3[3];
+    size_t gDim3[3];
     gDim3[0] = srcSize.width;
     gDim3[1] = srcSize.height;
     gDim3[2] = channel;
-    
+
     cl_kernel_implementer(theQueue, gDim3, NULL/*Local*/, theProgram, theKernel);
 
     return RPP_SUCCESS;
