+++ src/imgaug/cl/cl_declarations.hpp
@@ -8,7 +8,7 @@ cl_kernel_initializer ( cl_command_queue theQueue,
                         cl_program& theProgram, cl_kernel& theKernel);
 
 cl_int
-cl_kernel_implementer (cl_command_queue theHandle, unsigned int* gDim3, unsigned int* lDim3, cl_program& theProgram,
+cl_kernel_implementer (cl_command_queue theHandle, size_t* gDim3, size_t* lDim3, cl_program& theProgram,
                         cl_kernel& theKernel  );
 
 //===== Internal CL functions
@@ -37,12 +37,12 @@ cl_flip(cl_mem srcPtr, RppiSize srcSize,
                 RppiChnFormat chnFormat, unsigned int channel,
                 cl_command_queue theQueue);
 
-RppStatus 
+RppStatus
 cl_convert_rgb2hsv(cl_mem srcPtr, RppiSize srcSize,
                 cl_mem dstPtr,RppiChnFormat chnFormat, unsigned int chanel,
                 cl_command_queue theQueue);
 
-RppStatus 
+RppStatus
 cl_convert_hsv2rgb(cl_mem srcPtr, RppiSize srcSize,
                 cl_mem dstPtr,RppiChnFormat chnFormat, unsigned int chanel,
                 cl_command_queue theQueue);
@@ -56,5 +56,5 @@ RppStatus
 cl_hue_saturation_hsv( cl_mem srcPtr, RppiSize srcSize,
                 cl_mem dstPtr, Rpp32f hue, Rpp32f Saturation,
                 RppiChnFormat chnFormat, unsigned int channel, cl_command_queue theQueue);
-                
+
 #endif //RPP_CL_IMGAUG_DECLATAIONS_H
